+++ src/__init__.py
@@ -25,7 +25,7 @@ import requests
 from plyer import notification
 
 
-__version__ = "2021.01.15.11"
+__version__ = "2022.03.03.12"
 
 
 def sout(msg, clr='white'):