+++ docs/CHANGES.md
@@ -2,6 +2,14 @@
 
 ------
 
+## [build20220303]
+
+### Features
+
+* logging using **eventScript**
+
+
+
 ## [build20210712]
 
 ### Features